Move ISimpleCraftingRecipe to the proper package.
